@@ -1,4 +1,4 @@
-import { Usuario } from "@interfaces/usuario";
+import { Usuario } from "@interfaces/usuario.interfece";
 export interface Token {
   'refresh-token': string;
   token: string;
