@@ -3,13 +3,12 @@ import { Routes } from '@angular/router';
 export default [
   {
     path: 'dashboard',
-    canActivate: [],
     loadComponent: () => import('./admin-layout/admin-layout.component'),
     children: [
-    //   {
-    //     path: '',
-    //     loadChildren: () => import('../modules'),
-    //   },
+      {
+        path: 'facturacion',
+        loadComponent: () => import('./facturacion/componentes/base-facturacion/base-facturacion.component'),
+      },
     ],
   },
 ] as Routes;
