@@ -3,7 +3,8 @@ import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { usuarioActionInit } from '@redux/actions/usuario.actions';
 import { tap } from 'rxjs/operators';
-//import { getCookie, setCookie } from 'typescript-cookie';
+import { environment } from 'src/environment/environment';
+import { setCookie } from 'typescript-cookie';
 
 @Injectable()
 export class UsuarioEffects {
@@ -13,21 +14,21 @@ export class UsuarioEffects {
       this.actions$.pipe(
         ofType(usuarioActionInit),
         tap((action) => {
-          // let calcularTresHoras = new Date(
-          //   new Date().getTime() + 3 * 60 * 60 * 1000
-          // );
-          // if (environment.production) {
-          //   setCookie('usuario', JSON.stringify(action.usuario), {
-          //     expires: calcularTresHoras,
-          //     path: '/',
-          //     domain: environment.dominioApp,
-          //   });
-          // } else {
-          //   setCookie('usuario', JSON.stringify(action.usuario), {
-          //     expires: calcularTresHoras,
-          //     path: '/',
-          //   });
-          // }
+          let calcularTresHoras = new Date(
+            new Date().getTime() + 3 * 60 * 60 * 1000
+          );
+          if (environment.production) {
+            setCookie('usuario', JSON.stringify(action.usuario), {
+              expires: calcularTresHoras,
+              path: '/',
+              domain: environment.dominioApp,
+            });
+          } else {
+            setCookie('usuario', JSON.stringify(action.usuario), {
+              expires: calcularTresHoras,
+              path: '/',
+            });
+          }
         })
       ),
     { dispatch: false }
